+++ src/iterator/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod contiguous;
 pub mod pixel;
+pub mod raw;
 
 use crate::{
     draw_target::DrawTarget, geometry::Point, pixelcolor::PixelColor, primitives::Rectangle, Pixel,